requirements.txt
streamlit
spacy
transformers
wordcloud
matplotlib
pandas
textblob
scikit-learn
duckduckgo-search
newspaper3k
